+++ __tests__/all-rules/.eslintrc.js
@@ -3,6 +3,6 @@ const path = require("path");
 module.exports = {
   extends: [path.resolve(__dirname, "../../lib/config/all.js")],
   rules: {
-    "node/no-unsupported-features/es-syntax": "off"
-  }
+    "node/no-unsupported-features/es-syntax": "off",
+  },
 };