+++ __tests__/fixtures/react/App.jsx
@@ -9,7 +9,7 @@ class App extends React.Component {
     super(props);
 
     this.state = {
-      username: "hesmaili"
+      username: "hesmaili",
     };
   }