+++ __tests__/index.test.js
@@ -9,14 +9,14 @@ function isObject(obj) {
 }
 
 function dash2CamelCase(input) {
-  return input.replace(/-([a-z])/gu, found => found[1].toUpperCase());
+  return input.replace(/-([a-z])/gu, (found) => found[1].toUpperCase());
 }
 
-test("should the config is an object", t => {
+test("should the config is an object", (t) => {
   t.true(isObject(configs), "configs is an object");
 });
 
-test("should the `eslint` and the `eslint` CLI present", t => {
+test("should the `eslint` and the `eslint` CLI present", (t) => {
   t.true(isObject(eslint), "the `eslint` is present");
   t.true(
     typeof eslint.CLIEngine === "function",
@@ -24,7 +24,7 @@ test("should the `eslint` and the `eslint` CLI present", t => {
   );
 });
 
-test("should all configs are present in exports", t => {
+test("should all configs are present in exports", (t) => {
   const configDir = path.resolve(__dirname, "../lib/config");
   let files = [];
 
@@ -32,8 +32,8 @@ test("should all configs are present in exports", t => {
   files = fs.readdirSync(configDir);
 
   const actual = files
-    .filter(resource => resource !== ".eslintrc.js" && resource !== "rules")
-    .map(resource => dash2CamelCase(path.basename(resource, ".js")))
+    .filter((resource) => resource !== ".eslintrc.js" && resource !== "rules")
+    .map((resource) => dash2CamelCase(path.basename(resource, ".js")))
     .sort();
 
   const excepted = Object.keys(configs).sort();
@@ -41,13 +41,13 @@ test("should all configs are present in exports", t => {
   t.deepEqual(actual, excepted, "all configs are present in export");
 });
 
-test("should load the 'base' preset", t => {
+test("should load the 'base' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/base.js"] },
     rules: {
-      "no-undef": "error"
-    }
+      "no-undef": "error",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -55,11 +55,11 @@ test("should load the 'base' preset", t => {
   // t.is(configForFile.parser, require.resolve("babel-eslint"));
   t.deepEqual(configForFile.parserOptions, {
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
 
   const scriptReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/script.js")
+    path.resolve(__dirname, "./fixtures/script.js"),
   ]);
 
   t.is(scriptReport.results.length, 1);
@@ -67,7 +67,7 @@ test("should load the 'base' preset", t => {
   t.is(scriptReport.warningCount, 0);
 
   const moduleReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/module.js")
+    path.resolve(__dirname, "./fixtures/module.js"),
   ]);
 
   t.is(moduleReport.results.length, 1);
@@ -81,13 +81,13 @@ test("should load the 'base' preset", t => {
   );
 });
 
-test("should load the 'script' preset", t => {
+test("should load the 'script' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/script.js"] },
     rules: {
-      "no-undef": "error"
-    }
+      "no-undef": "error",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -95,16 +95,16 @@ test("should load the 'script' preset", t => {
   t.is(configForFile.parser, require.resolve("babel-eslint"));
   t.deepEqual(configForFile.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(configForFile.plugins.includes("import"));
   t.true(configForFile.plugins.includes("node"));
 
   const scriptReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/script.js")
+    path.resolve(__dirname, "./fixtures/script.js"),
   ]);
 
   t.is(scriptReport.results.length, 1);
@@ -112,7 +112,7 @@ test("should load the 'script' preset", t => {
   t.is(scriptReport.warningCount, 0);
 
   const moduleReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/module.js")
+    path.resolve(__dirname, "./fixtures/module.js"),
   ]);
 
   t.is(moduleReport.results.length, 1);
@@ -126,48 +126,13 @@ test("should load the 'script' preset", t => {
   );
 });
 
-test("should load the 'module' preset", t => {
+test("should load the 'module' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/module.js"] },
     rules: {
-      "no-undef": "error"
-    }
-  });
-
-  const configForFile = cli.getConfigForFile("myfile.js");
-
-  t.is(configForFile.parser, require.resolve("babel-eslint"));
-  t.deepEqual(configForFile.parserOptions, {
-    allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
-    sourceType: "module"
-  });
-
-  const scriptReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/script.js")
-  ]);
-
-  t.is(scriptReport.results.length, 1);
-  t.is(scriptReport.errorCount, 4);
-  t.is(scriptReport.warningCount, 0);
-
-  const moduleReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/module.js")
-  ]);
-
-  t.is(moduleReport.results.length, 1);
-  t.is(moduleReport.errorCount, 0);
-  t.is(moduleReport.warningCount, 0);
-});
-
-test("should load the 'dirty' preset", t => {
-  const cli = new eslint.CLIEngine({
-    useEslintrc: false,
-    baseConfig: { extends: ["./lib/config/dirty.js"] },
-    rules: {
-      "no-undef": "error"
-    }
+      "no-undef": "error",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -177,19 +142,18 @@ test("should load the 'dirty' preset", t => {
     allowImportExportEverywhere: true,
     ecmaVersion: 2020,
     sourceType: "module",
-    ecmaFeatures: { globalReturn: true }
   });
 
   const scriptReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/script.js")
+    path.resolve(__dirname, "./fixtures/script.js"),
   ]);
 
   t.is(scriptReport.results.length, 1);
-  t.is(scriptReport.errorCount, 0);
+  t.is(scriptReport.errorCount, 4);
   t.is(scriptReport.warningCount, 0);
 
   const moduleReport = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/module.js")
+    path.resolve(__dirname, "./fixtures/module.js"),
   ]);
 
   t.is(moduleReport.results.length, 1);
@@ -197,13 +161,49 @@ test("should load the 'dirty' preset", t => {
   t.is(moduleReport.warningCount, 0);
 });
 
-test("should load the 'ava' preset", t => {
+test("should load the 'dirty' preset", (t) => {
+  const cli = new eslint.CLIEngine({
+    useEslintrc: false,
+    baseConfig: { extends: ["./lib/config/dirty.js"] },
+    rules: {
+      "no-undef": "error",
+    },
+  });
+
+  const configForFile = cli.getConfigForFile("myfile.js");
+
+  t.is(configForFile.parser, require.resolve("babel-eslint"));
+  t.deepEqual(configForFile.parserOptions, {
+    allowImportExportEverywhere: true,
+    ecmaVersion: 2020,
+    sourceType: "module",
+    ecmaFeatures: { globalReturn: true },
+  });
+
+  const scriptReport = cli.executeOnFiles([
+    path.resolve(__dirname, "./fixtures/script.js"),
+  ]);
+
+  t.is(scriptReport.results.length, 1);
+  t.is(scriptReport.errorCount, 0);
+  t.is(scriptReport.warningCount, 0);
+
+  const moduleReport = cli.executeOnFiles([
+    path.resolve(__dirname, "./fixtures/module.js"),
+  ]);
+
+  t.is(moduleReport.results.length, 1);
+  t.is(moduleReport.errorCount, 0);
+  t.is(moduleReport.warningCount, 0);
+});
+
+test("should load the 'ava' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/module.js", "./lib/config/ava.js"] },
     rules: {
-      "ava/no-ignored-test-files": "off"
-    }
+      "ava/no-ignored-test-files": "off",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -211,13 +211,13 @@ test("should load the 'ava' preset", t => {
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
 
   t.true(configForFile.plugins.includes("ava"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/ava.mjs")
+    path.resolve(__dirname, "./fixtures/ava.mjs"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -225,19 +225,19 @@ test("should load the 'ava' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'esnext' preset", t => {
+test("should load the 'esnext' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
       extends: [
         "./lib/config/module.js",
         "./lib/config/browser.js",
-        "./lib/config/esnext.js"
-      ]
+        "./lib/config/esnext.js",
+      ],
     },
     rules: {
-      "max-classes-per-file": "off"
-    }
+      "max-classes-per-file": "off",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -245,7 +245,7 @@ test("should load the 'esnext' preset", t => {
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
   t.true(configForFile.env.es2020);
   t.true(configForFile.env.browser);
@@ -255,7 +255,7 @@ test("should load the 'esnext' preset", t => {
   t.true(configForFile.plugins.includes("unicorn"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/good.js")
+    path.resolve(__dirname, "./fixtures/good.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -263,13 +263,13 @@ test("should load the 'esnext' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'lodash' preset", t => {
+test("should load the 'lodash' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/module.js", "./lib/config/lodash"] },
     rules: {
-      "import/no-extraneous-dependencies": "off"
-    }
+      "import/no-extraneous-dependencies": "off",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -277,13 +277,13 @@ test("should load the 'lodash' preset", t => {
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
 
   t.true(configForFile.plugins.includes("lodash"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/lodash.js")
+    path.resolve(__dirname, "./fixtures/lodash.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -291,29 +291,29 @@ test("should load the 'lodash' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'node' preset", t => {
+test("should load the 'node' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
       extends: [
         "./lib/config/script.js",
         "./lib/config/esnext.js",
-        "./lib/config/node.js"
-      ]
+        "./lib/config/node.js",
+      ],
     },
     rules: {
-      "no-console": "off"
-    }
+      "no-console": "off",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
 
   t.deepEqual(configForFile.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(configForFile.env.node);
   t.false(Boolean(configForFile.env.browser));
@@ -321,7 +321,7 @@ test("should load the 'node' preset", t => {
   t.true(configForFile.plugins.includes("node"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/node.js")
+    path.resolve(__dirname, "./fixtures/node.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -329,32 +329,32 @@ test("should load the 'node' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'browser' preset", t => {
+test("should load the 'browser' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
       extends: [
         "./lib/config/script.js",
         "./lib/config/browser.js",
-        "./lib/config/esnext.js"
-      ]
-    }
+        "./lib/config/esnext.js",
+      ],
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
 
   t.deepEqual(configForFile.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.false(Boolean(configForFile.env.node));
   t.true(configForFile.env.browser);
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/browser.js")
+    path.resolve(__dirname, "./fixtures/browser.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -362,7 +362,7 @@ test("should load the 'browser' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load 'node' and 'browser' presets", t => {
+test("should load 'node' and 'browser' presets", (t) => {
   const cliNodeAndBrowser = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
@@ -370,13 +370,13 @@ test("should load 'node' and 'browser' presets", t => {
         "./lib/config/script.js",
         "./lib/config/esnext.js",
         "./lib/config/node.js",
-        "./lib/config/browser.js"
-      ]
+        "./lib/config/browser.js",
+      ],
     },
     rules: {
       "no-console": "off",
-      "import/no-nodejs-modules": "off"
-    }
+      "import/no-nodejs-modules": "off",
+    },
   });
 
   const configForFileNodeAndBrowser = cliNodeAndBrowser.getConfigForFile(
@@ -385,10 +385,10 @@ test("should load 'node' and 'browser' presets", t => {
 
   t.deepEqual(configForFileNodeAndBrowser.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(configForFileNodeAndBrowser.env.node);
   t.true(configForFileNodeAndBrowser.env.browser);
@@ -405,7 +405,7 @@ test("should load 'node' and 'browser' presets", t => {
 
   const reportNodeAndBrowser = cliNodeAndBrowser.executeOnFiles([
     path.resolve(__dirname, "./fixtures/node.js"),
-    path.resolve(__dirname, "./fixtures/browser.js")
+    path.resolve(__dirname, "./fixtures/browser.js"),
   ]);
 
   t.is(
@@ -423,13 +423,13 @@ test("should load 'node' and 'browser' presets", t => {
         "./lib/config/script.js",
         "./lib/config/esnext.js",
         "./lib/config/browser.js",
-        "./lib/config/node.js"
-      ]
+        "./lib/config/node.js",
+      ],
     },
     rules: {
       "no-console": "off",
-      "import/no-nodejs-modules": "off"
-    }
+      "import/no-nodejs-modules": "off",
+    },
   });
 
   const configForFileBrowserAndNode = cliBrowserAndNode.getConfigForFile(
@@ -438,10 +438,10 @@ test("should load 'node' and 'browser' presets", t => {
 
   t.deepEqual(configForFileBrowserAndNode.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(configForFileBrowserAndNode.env.node);
   t.true(configForFileBrowserAndNode.env.browser);
@@ -458,7 +458,7 @@ test("should load 'node' and 'browser' presets", t => {
 
   const reportBrowserAndNode = cliBrowserAndNode.executeOnFiles([
     path.resolve(__dirname, "./fixtures/node.js"),
-    path.resolve(__dirname, "./fixtures/browser.js")
+    path.resolve(__dirname, "./fixtures/browser.js"),
   ]);
 
   t.is(
@@ -476,23 +476,23 @@ test("should load 'node' and 'browser' presets", t => {
         "./lib/config/script.js",
         "./lib/config/browser.js",
         "./lib/config/node.js",
-        "./lib/config/esnext.js"
-      ]
+        "./lib/config/esnext.js",
+      ],
     },
     rules: {
       "no-console": "off",
-      "import/no-nodejs-modules": "off"
-    }
+      "import/no-nodejs-modules": "off",
+    },
   });
 
   const configForFileESNextLast = cliESNextLast.getConfigForFile("myfile.js");
 
   t.deepEqual(configForFileESNextLast.parserOptions, {
     ecmaFeatures: {
-      globalReturn: true
+      globalReturn: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(configForFileESNextLast.env.node);
   t.true(configForFileESNextLast.env.browser);
@@ -508,7 +508,7 @@ test("should load 'node' and 'browser' presets", t => {
 
   const reportESNextLast = cliESNextLast.executeOnFiles([
     path.resolve(__dirname, "./fixtures/node.js"),
-    path.resolve(__dirname, "./fixtures/browser.js")
+    path.resolve(__dirname, "./fixtures/browser.js"),
   ]);
 
   t.is(reportESNextLast.results.length, 2, "eslint report with one results");
@@ -516,7 +516,7 @@ test("should load 'node' and 'browser' presets", t => {
   t.is(reportESNextLast.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'react' preset", t => {
+test("should load the 'react' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
@@ -524,13 +524,13 @@ test("should load the 'react' preset", t => {
         "./lib/config/module.js",
         "./lib/config/browser.js",
         "./lib/config/esnext.js",
-        "./lib/config/react.js"
-      ]
+        "./lib/config/react.js",
+      ],
     },
     rules: {
       "no-console": "off",
-      "react/prefer-stateless-function": "off"
-    }
+      "react/prefer-stateless-function": "off",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -538,17 +538,17 @@ test("should load the 'react' preset", t => {
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
     ecmaFeatures: {
-      jsx: true
+      jsx: true,
     },
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
   t.true(configForFile.env.browser);
   t.true(configForFile.plugins.includes("react"));
   t.true(configForFile.plugins.includes("jsx-a11y"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/react/**/*.jsx")
+    path.resolve(__dirname, "./fixtures/react/**/*.jsx"),
   ]);
 
   t.is(report.results.length, 3, "eslint report with one results");
@@ -556,13 +556,13 @@ test("should load the 'react' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'html' preset", t => {
+test("should load the 'html' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/html.js"] },
     rules: {
-      "no-alert": "error"
-    }
+      "no-alert": "error",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -579,7 +579,7 @@ test("should load the 'html' preset", t => {
     var i = 1;
   </script>
  </head>
- <body> 
+ <body>
   <div id="welcome"></div>
  </body>
 </html>`,
@@ -600,7 +600,7 @@ test("should load the 'html' preset", t => {
   alert('test');
   </script>
  </head>
- <body> 
+ <body>
   <div id="welcome"></div>
  </body>
 </html>`,
@@ -612,12 +612,12 @@ test("should load the 'html' preset", t => {
   t.is(invalidReport.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'jest' preset", t => {
+test("should load the 'jest' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
-      extends: ["./lib/config/module", "./lib/config/jest"]
-    }
+      extends: ["./lib/config/module", "./lib/config/jest"],
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -626,7 +626,7 @@ test("should load the 'jest' preset", t => {
   t.true(configForFile.plugins.includes("jest"));
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/jest.js")
+    path.resolve(__dirname, "./fixtures/jest.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -634,21 +634,21 @@ test("should load the 'jest' preset", t => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
-test("should load the 'markdown' preset", t => {
+test("should load the 'markdown' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: { extends: ["./lib/config/markdown"] },
     rules: {
-      "no-alert": "error"
-    }
+      "no-alert": "error",
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
 
   t.deepEqual(configForFile.parserOptions, {
     ecmaFeatures: {
-      impliedStrict: true
-    }
+      impliedStrict: true,
+    },
   });
   t.true(configForFile.plugins.includes("markdown"));
 
@@ -682,12 +682,12 @@ alert("test");
       extends: [
         "./lib/config/script",
         "./lib/config/node",
-        "./lib/config/markdown"
-      ]
+        "./lib/config/markdown",
+      ],
     },
     rules: {
-      "import/no-unresolved": "off"
-    }
+      "import/no-unresolved": "off",
+    },
   });
 
   const scriptConfigForFile = scriptCli.getConfigForFile("myfile.js");
@@ -695,10 +695,10 @@ alert("test");
   t.deepEqual(scriptConfigForFile.parserOptions, {
     ecmaFeatures: {
       globalReturn: true,
-      impliedStrict: true
+      impliedStrict: true,
     },
     ecmaVersion: 2020,
-    sourceType: "script"
+    sourceType: "script",
   });
   t.true(scriptConfigForFile.plugins.includes("markdown"));
 
@@ -740,15 +740,15 @@ console.log(test);
       extends: [
         "./lib/config/module",
         "./lib/config/node",
-        "./lib/config/markdown"
-      ]
+        "./lib/config/markdown",
+      ],
     },
     rules: {
       "import/no-unresolved": "off",
       "import/no-extraneous-dependencies": "off",
       "node/no-unsupported-features/es-syntax": "off",
-      "no-undef": "error"
-    }
+      "no-undef": "error",
+    },
   });
 
   const moduleConfigForFile = moduleCli.getConfigForFile("myfile.js");
@@ -756,10 +756,10 @@ console.log(test);
   t.deepEqual(moduleConfigForFile.parserOptions, {
     allowImportExportEverywhere: true,
     ecmaFeatures: {
-      impliedStrict: true
+      impliedStrict: true,
     },
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
   t.true(moduleConfigForFile.plugins.includes("markdown"));
 
@@ -797,8 +797,8 @@ console.log(test);
       extends: [
         "./lib/config/dirty",
         "./lib/config/node",
-        "./lib/config/markdown"
-      ]
+        "./lib/config/markdown",
+      ],
     },
     rules: {
       "no-undef": "error",
@@ -806,8 +806,8 @@ console.log(test);
       "import/no-unresolved": "off",
       "import/no-extraneous-dependencies": "off",
       "import/extensions": "off",
-      "node/no-unsupported-features/es-syntax": "off"
-    }
+      "node/no-unsupported-features/es-syntax": "off",
+    },
   });
 
   const dirtyConfigForFile = dirtyCli.getConfigForFile("myfile.js");
@@ -816,10 +816,10 @@ console.log(test);
     allowImportExportEverywhere: true,
     ecmaFeatures: {
       globalReturn: true,
-      impliedStrict: true
+      impliedStrict: true,
     },
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
   t.true(dirtyConfigForFile.plugins.includes("markdown"));
 
@@ -860,16 +860,16 @@ var qwerty = 1;
   t.true(invalidDirtyReport.results[0].messages[0].ruleId === "no-var");
 });
 
-test("integration tests for unused eslint comments", t => {
+test("integration tests for unused eslint comments", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
-      extends: ["./lib/config/module.js", "./lib/config/esnext.js"]
-    }
+      extends: ["./lib/config/module.js", "./lib/config/esnext.js"],
+    },
   });
 
   const report = cli.executeOnFiles([
-    path.resolve(__dirname, "./fixtures/unused-eslint-directives.js")
+    path.resolve(__dirname, "./fixtures/unused-eslint-directives.js"),
   ]);
 
   t.is(report.results.length, 1, "eslint report with one results");
@@ -888,7 +888,7 @@ test("integration tests for unused eslint comments", t => {
   );
 });
 
-test("should load the 'all' preset", t => {
+test("should load the 'all' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
     baseConfig: {
@@ -902,9 +902,9 @@ test("should load the 'all' preset", t => {
         "./lib/config/react.js",
         "./lib/config/jest.js",
         "./lib/config/ava.js",
-        "./lib/config/lodash.js"
-      ]
-    }
+        "./lib/config/lodash.js",
+      ],
+    },
   });
 
   const configForFile = cli.getConfigForFile("myfile.js");
@@ -914,10 +914,10 @@ test("should load the 'all' preset", t => {
     ecmaFeatures: {
       globalReturn: true,
       impliedStrict: true,
-      jsx: true
+      jsx: true,
     },
     ecmaVersion: 2020,
-    sourceType: "module"
+    sourceType: "module",
   });
   t.true(configForFile.plugins.includes("ava"));
   t.true(configForFile.plugins.includes("import"));